Started on the implementation of an automatic unlock system for booked lockers after 15hours as a max allocated booking time.
Release_expired_lockers_logic now fully works and it has been tested, added the log_reserved_action as a log for when the locker was first reserved so the logic can see how much time the user has of remaining time before it gets automatically unreserved\unlocked and cleared of its user.

changed all timings to correctly check for locker statuses every 10 minutes and set a timer for 15 hours until actual locker unlock.

finally, added the background process release_expired_loop to start every 10 minutes to check if a locker booking has reached 15 hours.(it is currently set to 10 seconds for testing purposes)

made sure to keep all instances of statuses to start with a capital letter(should change this to ilike later so it doesnt get case-sensitive).

added log_action function for any logging, it is not normally used but was used for testing and it will be kept for further testing.

tried to fix get_total_users