cleaned up main.py a little.
added some API-calls for temporary_unlock_endpoint,
manual_release_locker_endpoint and two statistic endpoints that currently is incomplete.

added the function manual_release_locker that is used by the standardUser to end their day, clear their id from the locker and make the locker ready for the next user.

added the function temporary_unlock which lets the StandardUser open their locker without u-reserving or make the locker say that it is "Ledig"\ ready for another user by not being marked as occupied anymore.

added the function lock_locker_after_use to make sure there is an in-use function for logging that the locker is locked again after beeing opened by temporary_unlock.

tried to fix up some statistic.py functions, but this seems like a tast for later so im just leaving the current changes for now, will come back to this later. there is much that is not currently working as intended and this file needs a cleanup.